pyproject.toml
[project]
name = "mcp-project-one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aisuite[all]>=0.1.6",
    "arxiv==2.2.0",
    "fastmcp==2.5.2",
    "flask==3.1.1",
    "google-ai-generativelanguage==0.6.15",
    "google-genai==1.18.0",
    "google-generativeai==0.8.5",
    "httpx==0.28.1",
    "huggingface-hub==0.32.3",
    "ipykernel==6.29.5",
    "ipython==8.36.0",
    "markdown==3.8",
    "markupsafe==3.0.2",
    "mcp[cli]==1.9.2",
    "python-dotenv==1.1.0",
    "streamlit>=1.45.1",
    "vertexai==1.71.1",
    "youtube-search-python>=1.6.6",
    "yt-dlp==2025.5.22",
]
